Moved saving to left toolbar
